requirements.txt
full-fred==0.0.9a3
requests==2.26.0
google-cloud-firestore==2.3.4
